use nvme1n1

+++ install.sh
@@ -9,12 +9,15 @@ DRIVER_BASE_URL=https://us.download.nvidia.com/tesla
 DRIVER_VERSION=525.85.12
 
 # 內建的nvme1
-sudo file -s /dev/nvme1n1
-sudo mkfs -t xfs /dev/nvme1n1
-sudo mkdir /data
-sudo mount /dev/nvme1n1 /data
-sudo chmod 777 /data
-echo '/dev/nvme1n1 /data ext4 discard,defaults,nofail 0 2' | sudo tee -a /etc/fstab
+if [ ! -d "/data" ]
+then
+    sudo file -s /dev/nvme1n1
+    sudo mkfs -t xfs /dev/nvme1n1
+    sudo mkdir /data
+    sudo mount /dev/nvme1n1 /data
+    sudo chown ${USER}.${USER} /data
+    echo '/dev/nvme1n1 /data ext4 discard,defaults,nofail 0 2' | sudo tee -a /etc/fstab
+fi
 
 # 安裝系統套件
 sudo apt update
@@ -24,9 +27,14 @@ sudo apt install -y wget curl aria2 \
 
 # 安裝顯示卡驅動
 # https://www.nvidia.com/Download/Find.aspx
-cd /tmp
-curl -fSsl -O $DRIVER_BASE_URL/$DRIVER_VERSION/NVIDIA-Linux-$ARCH-$DRIVER_VERSION.run
-sudo sh NVIDIA-Linux-$ARCH-$DRIVER_VERSION.run -s
+if [ ! -f "/root/.nvdirve_installed" ]
+then
+    cd /tmp
+    curl -fSsl -O $DRIVER_BASE_URL/$DRIVER_VERSION/NVIDIA-Linux-$ARCH-$DRIVER_VERSION.run
+    sudo sh NVIDIA-Linux-$ARCH-$DRIVER_VERSION.run -s
+    rm NVIDIA-Linux-$ARCH-$DRIVER_VERSION.run
+    sudo touch /root/.nvdirve_installed
+fi
 
 # 安裝stable-diffusion-webui
 cd ${HOME_DIR}
